Factor some CardanoClient queries to drop hydra-cluster dependency in hydra-tui
